# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agency Group — Luxury Real Estate Portugal API",
    "description": "Search luxury properties in Portugal (Lisbon, Cascais, Comporta, Porto, Algarve, Madeira). Get market data, valuations, NHR tax info. AMI 22506.",
    "version": "1.0.0",
    "contact": {
      "name": "Agency Group",
      "url": "https://www.agencygroup.pt",
      "email": "geral@agencygroup.pt"
    }
  },
  "servers": [
    { "url": "https://www.agencygroup.pt" }
  ],
  "paths": {
    "/api/gpt/properties": {
      "get": {
        "operationId": "searchProperties",
        "summary": "Search luxury properties in Portugal",
        "description": "Search for luxury properties by zone, type, price range, bedrooms. Returns available properties with price, area, features.",
        "parameters": [
          { "name": "zona", "in": "query", "description": "Zone/city: Lisboa, Cascais, Comporta, Porto, Algarve, Madeira, Sintra, Ericeira", "schema": { "type": "string" } },
          { "name": "tipo", "in": "query", "description": "Property type: Apartamento, Moradia, Villa, Quinta, Penthouse, Herdade", "schema": { "type": "string" } },
          { "name": "precoMin", "in": "query", "description": "Minimum price in EUR", "schema": { "type": "integer" } },
          { "name": "precoMax", "in": "query", "description": "Maximum price in EUR", "schema": { "type": "integer" } },
          { "name": "quartos", "in": "query", "description": "Minimum number of bedrooms", "schema": { "type": "integer" } },
          { "name": "lang", "in": "query", "description": "Response language: pt, en, fr", "schema": { "type": "string", "default": "en" } }
        ],
        "responses": {
          "200": {
            "description": "List of matching properties",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "properties": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "zone": { "type": "string" },
                          "type": { "type": "string" },
                          "price": { "type": "integer" },
                          "area": { "type": "integer" },
                          "bedrooms": { "type": "integer" },
                          "features": { "type": "array", "items": { "type": "string" } },
                          "url": { "type": "string" },
                          "pricePerSqm": { "type": "integer" }
                        }
                      }
                    },
                    "total": { "type": "integer" },
                    "marketNote": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}